Class for parsing based on joint keys
Add a class that takes keys from which to extract values from a dict and surrounding strings. Allow multiple instances to be grouped together to parse only the first key found in the group. Also allow multiple keys to be grouped in a single instance so that the keys will be parsed only if they are all present. Use this class for the Scopus journal extractor, which parses keys sequentially based on these types of key combinations.