Replace database extractors with YAML files
Each database extractor has been defined by a dictionary. To allow new extractors to be defined outside of code, parse extractors as YAML files instead of Python dicts. When parsing the YAML files, translate specific classes defined in a string to the given class to provide access to the `ExtractKeys` Enums and  `JointKeyExtractor` class instances.