Parse each database file as it is selected
The GUI has silently taken all database files as input and provided a button to parse all the files and find overlaps. Separate out the parsing tasks, performing database extraction on each file as soon as it is loaded. Provide a dropdown to choose the extractor, including an "Auto" option to detect the extractor based on the file name. Add an "Import" button to perform the load and extraction, plus a table for the given entry. Add a separate table for each of the main database inputs and load the tables based on the input file.