Use Pandas data frame to process database
Take a data frame rather than handling the CSV file directly to allow importing the CSV file before the processing. Use the Pandas rather than the native CSV reader to handle file encoding detection, which obviates the need for parsing the byte-order mark for the signature UTF-8 (`utf-8-sig`) encoding of the Embase and SCOPUS files.