Fix finding the database name from paths without an underscore
The database extraction method parses the given path to find the database name on the assumption that the the first segment before an underscore is the database name. For names without any underscore, this segment contains the extension and thus does not match any existing database names. Remove the extension to allow paths without any underscores.