More flexibility in finding database name
The database name has been found based on a match to the first segment of the filename, but this first segment has also been used to find the extractor, which may not match any known database name. To provide more flexibility in finding the database name, only require this first segment to start with a known database to consider the database a match.