Find database name based on chosen extractor rather than on input path
The database name has been determined based on the file path, but the extractor path parameter has been used as a way to choose the extractor the appropriate database even if the file name does not match or identify a given database. To allow any arbitrary filename to be given if an extractor path is given, use the extractor path instead to determine the database name. The filename may still be used indirectly to find the database name by determining the chosen extractor path, such as through the current behavior in the CLI.