Allow more flexibility for database sheet names
The sheet name is used to identify the appropriate extractor for the given sheet, but the `findOverlaps` function has only looked for sheets whose names match the database names constant. Rather than looking for an exact match, only require that the sheet names start with a given database to consider the sheet a match for that database.