Fix finding extractor paths outside of the app root folder
Use the app's directory to construct the path to the extractor folder rather than using a relative path assumed to be in the app root folder. Limit the names of extractors in the GUI to the filenames rather than the full path since this path may be long and widen the window.