Fix expanding window when saving tables
The status bar displays the output message from saving tables to files, which can be a long string if paths are long. The status bar automatically expands to accommodate the full width of this string, which can dramatically increase the size of the window. Reduce the string size by only displaying the path of the extracted overlap file, split into its basename and the folder that contains it and the filtered files.