Refactor overlap functions into `DbOverlapper` class
Similarly to the `DbExtractor` class, the overlap functionality requires use of "global" dictionaries that are passed to and from each function. Move these overlap functions into a new class to encapsulate these global dictionaries. Create a mixin class that both the database overlap and extractor classes extend to set up these dictionaries. Since the match finder is similar between these two classes, generalize it through the mixin class.

To avoid circular imports, move the arg parser and main entry point to the extractor module. This move fits the functionality of the entry point to access `DbExtractor` rather than the overlapper.