Encapsulate spreadsheets in separate class
Group array and table adapter into Traits-enabled class, which encapsulates the related variables and also allows the Array Traits to be reassigned from a sequence containing the object rather than accessing the variable directly. This change appears to require the arrays to be accessed through the `object.<var-name>.<attr>` syntax. In turn, this access names tabs by the attribute and requires the handler to rename tabs after setting up the GUI elements. Assume that the last tab is the "overlaps" tab.