Generalize merging CSV/TSV files from a directory
Accept a single path in the CSV merge function. If the path is a directory, merge its entire contents. If not, simply read the single file. Replace NaNs with empty strings in the combined data frame since it will contains NaNs for any missing columns in a given file despite removing NaNs when reading each input file.