Get current GAS spreadsheet from user properties if active spreadsheet not available
The active spreadsheet must be set during each function call when the GAS app is run not run as an Add-on. To retain a reference to the spreadsheet, provide a function to store the spreadsheet URL in user properties. Retrieve a spreadsheet from this URL if an active spreadsheet is not available.

Although the API docs say that the `spreadsheets.currentonly` spreadsheet scope is valid for the `openByUrl` all, it gives an exception saying that the full `spreadsheets` scope is required. Change the spreadsheets scope to this larger permission.