Option to run the GAS GUI as a web app
The GAS GUI has been run as an Add-on attached to a given spreadsheet. As an alternative, provide the option to the run the GUI as a web app. Reuse the sidebar in a wrapper to reproduce the same controls, adding a URL input field to specify the spreadsheet to work on. Load this spreadsheet into a separate iframe shown next to the sidebar. Add a web app deployment accessible to anyone as the user.