Upload and import citation list files directly in web app
The import file function is disabled when a Google spreadsheet is loaded in an iframe. To provide this file import functionality in the web app, add a file upload control and parse the file contents directly into a new sheet. This control replaces the sheet setup control since each file is uploaded into a new sheet named by the selected database source.