Fix importing TSV files without `.tsv` extension
File delimiters are determined by extension since auto-detection in Pandas has not worked reliably in our tests for TSV files. The delimiter has been set to tabs only for files ending in a `.tsv` extension, but other files such as `.txt` may also use tabs as delimiters. To extend TSV support to files without a `.tsv` extension, fall back to reading with tab delimiters if comma-delimited reading fails. This change continues to assume that only tab and comma delimiters are used but allows tabs to be used without the appropriate extension and also supports incorrect extensions for CSV and TSV files.