Support opening URL hyperlinks in an external browser
The TraitsUI `HTMLEditor` falls back to using a `QTextBrowser` when `QtWebEngine` is not installed, as is the case for our environments, which do not install this package. While the fallback editor displays HTML formatted text, including clickable links, the links do not open the page in an external web browser by default. Update the `QTextBrowser` widget in the UI handler to open these links in an external browser.